refactor(ios): dedupe haversine calc into Coordinate.distanceMetres (tc-6he3x.7)
WatchZone.haversineDistance and WatchZoneBoundary.haversineMetres were
identical great-circle calculations with the same earth radius,
flagged by CodeRabbit review as a maintenance nitpick on PR #1048.
Moved to Coordinate.distanceMetres(to:), with a raw-doubles overload
for WatchZoneBoundary.enclosingRadiusMetres's centroid (a derived
point, not a validated Coordinate).
